+++ ReleaseNotes.md
@@ -1,2 +1,5 @@
+# Prismic .NET Standard Release Notes
 
+## v1.4.0
+Support boolean fragments